Split V1
Split skinnerbox.py into its individual components. Also updated main.css to be consistent throughout webpages. Also changed trialpage.html to trialsettingspage.html.
