+++ app/__init__.py
@@ -1,0 +1,6 @@
+# app/__init__.py
+from flask import Flask
+
+app = Flask(__name__, template_folder='../templates', static_folder='../static')
+
+from app import routes
