+++ templates/homepage.html
@@ -3,11 +3,12 @@
 {% block head %}
 {% endblock %}
 
-{% block body %} 
+{% block body %}
 
 <div class="container">
     <header>
         <h1>Welcome to the Skinnerbox App</h1>
+        
     </header>
     
     <section class="get-started">
