+++ templates/trialsettingspage.html
@@ -6,11 +6,8 @@
 
 {% block body %}
 
-
-
 <div class="container ">
-    <h1>Trial Page</h1>
-    <h2>Trial Settings</h2>
+    <h1>Trial Settings</h1>
     <form action="{{ url_for('update_trial_settings') }}" method="post">
         <div class="form-group">
             <label for="duration">Trial Duration (minutes)</label>
